2b

+++ HW5/2a/RNN_test.py
@@ -107,7 +107,7 @@ for epoch in range(no_of_epochs):
                 x_input[j,:] = x[start_index:(start_index+sequence_length)]
         y_input = y_test[I_permutation[i:i+ln]]
 
-        data = Variable(torch.LongTensor(x_input)).cuda()
+        data = Variable(torch.FloatTensor(x_input)).cuda()
         target = Variable(torch.FloatTensor(y_input)).cuda()
 
         # with torch.no_grad():
